+++ OpenWeatherMap.NetClient/Apis/Forecast5DaysApi.cs
@@ -1,27 +1,29 @@
 ﻿using OpenWeatherMap.NetClient.Extensions;
 using OpenWeatherMap.NetClient.Models;
 using OpenWeatherMap.NetClient.RestApis.Clients;
-using OpenWeatherMap.NetClient.RestApis.Responses;
-using Refit;
-using ApiException = OpenWeatherMap.NetClient.Exceptions.ApiException;
 
 namespace OpenWeatherMap.NetClient.Apis;
 
 /// <summary>
 /// Implementation of <see cref="IForecast5DaysApi"/>
 /// </summary>
-public sealed class Forecast5DaysApi : AbstractApiImplBase, IForecast5DaysApi
+public sealed class Forecast5DaysApi : IForecast5DaysApi
 {
+  private const string BaseUrl = "https://api.openweathermap.org";
+
   private readonly string _apiKey;
+  private readonly string _language;
 
-  private readonly IForecast5DaysApiClient _forecastApi;
-  private readonly IGeocodingApiClient _geoApi;
+  private readonly RestClient<IForecast5DaysApiClient> _forecastApi;
+  private readonly RestClient<IGeocodingApiClient> _geoApi;
 
-  internal Forecast5DaysApi(string apiKey, IOpenWeatherMapOptions? options) : base(options)
+  internal Forecast5DaysApi(string apiKey, IOpenWeatherMapOptions? options)
   {
     _apiKey = apiKey;
-    _forecastApi = RestService.For<IForecast5DaysApiClient>(BaseUrl);
-    _geoApi = RestService.For<IGeocodingApiClient>(BaseUrl);
+    _language = (options?.Culture ?? OpenWeatherMapOptions.Defaults.Culture).TwoLetterISOLanguageName;
+
+    _forecastApi = new RestClient<IForecast5DaysApiClient>(BaseUrl, options);
+    _geoApi = new RestClient<IGeocodingApiClient>(BaseUrl, options);
   }
 
   /// <inheritdoc />
@@ -29,51 +31,43 @@ public sealed class Forecast5DaysApi : AbstractApiImplBase, IForecast5DaysApi
   {
     if (query == null) throw new ArgumentNullException(nameof(query));
 
-    return await Cached(() => $"Forecast5DaysByName_{query}", async () =>
-    {
-      var response = await _geoApi.GeoCodeByLocationName(_apiKey, query, 1);
-      if (!response.IsSuccessStatusCode)
-      {
-        throw new ApiException(response.StatusCode, response.ReasonPhrase, response.Error);
-      }
+    var response = await _geoApi.Call(
+      api => api.GeoCodeByLocationName(_apiKey, query, 1),
+      () => $"GeoCodeByLocationName_{query}"
+    );
+    if (!response.Any()) return null;
 
-      if (response.Content == null || !response.Content.Any())
+    var geoCode = response.First();
+    return await _forecastApi.Call(async api =>
       {
-        return null;
-      }
-
-      var geoCode = response.Content.First();
-      return MapResponse(
-        await _forecastApi.Forecast(_apiKey, Language, geoCode.Latitude, geoCode.Longitude, limit)
-      );
-    });
+        var forecast = await api.Forecast(_apiKey, _language, geoCode.Latitude, geoCode.Longitude, limit);
+        return forecast.ToForecast();
+      },
+      () => $"Forecast5DaysByName_{query}"
+    );
   }
 
   /// <inheritdoc />
   public async Task<Forecast5Days?> GetByCoordinatesAsync(double lat, double lon, int limit = int.MaxValue)
   {
-    return await Cached(
-      () => $"Forecast5DaysByCoordinates_{lat}_{lon}",
-      async () => MapResponse(await _forecastApi.Forecast(_apiKey, Language, lat, lon, limit))
+    return await _forecastApi.Call(async api =>
+      {
+        var forecast = await api.Forecast(_apiKey, _language, lat, lon, limit);
+        return forecast.ToForecast();
+      },
+      () => $"Forecast5DaysByCoordinates_{lat}_{lon}"
     );
   }
 
   /// <inheritdoc />
   public async Task<Forecast5Days?> GetByCityIdAsync(int cityId, int limit = int.MaxValue)
   {
-    return await Cached(
-      () => $"Forecast5DaysByCityId_{cityId}",
-      async () => MapResponse(await _forecastApi.Forecast(_apiKey, Language, cityId, limit))
+    return await _forecastApi.Call(async api =>
+      {
+        var forecast = await api.Forecast(_apiKey, _language, cityId, limit);
+        return forecast.ToForecast();
+      },
+      () => $"Forecast5DaysByCityId_{cityId}"
     );
-  }
-
-  private static Forecast5Days? MapResponse(IApiResponse<ApiForecast5DaysResponse> response)
-  {
-    if (!response.IsSuccessStatusCode)
-    {
-      throw new ApiException(response.StatusCode, response.ReasonPhrase, response.Error);
-    }
-
-    return response.Content?.ToForecast();
   }
 }