+++ OpenWeatherMap.NetClient/Apis/GeocodingApi.cs
@@ -1,25 +1,24 @@
 ﻿using OpenWeatherMap.NetClient.Extensions;
 using OpenWeatherMap.NetClient.Models;
 using OpenWeatherMap.NetClient.RestApis.Clients;
-using OpenWeatherMap.NetClient.RestApis.Responses;
-using Refit;
-using ApiException = OpenWeatherMap.NetClient.Exceptions.ApiException;
 
 namespace OpenWeatherMap.NetClient.Apis;
 
 /// <summary>
 /// Implementation of <see cref="IGeocodingApi"/>
 /// </summary>
-public sealed class GeocodingApi : AbstractApiImplBase, IGeocodingApi
+public sealed class GeocodingApi : IGeocodingApi
 {
+  private const string BaseUrl = "https://api.openweathermap.org";
+
   private readonly string _apiKey;
 
-  private readonly IGeocodingApiClient _geoCodingApiClient;
+  private readonly RestClient<IGeocodingApiClient> _geoApi;
 
-  internal GeocodingApi(string apiKey, IOpenWeatherMapOptions? options) : base(options)
+  internal GeocodingApi(string apiKey, IOpenWeatherMapOptions? options)
   {
     _apiKey = apiKey;
-    _geoCodingApiClient = RestService.For<IGeocodingApiClient>(BaseUrl);
+    _geoApi = new RestClient<IGeocodingApiClient>(BaseUrl, options);
   }
 
   /// <inheritdoc />
@@ -27,8 +26,12 @@ public sealed class GeocodingApi : AbstractApiImplBase, IGeocodingApi
   {
     if (query == null) throw new ArgumentNullException(nameof(query));
 
-    return await Cached(() => $"GeoCode_{query}_{limit}", async () =>
-      MapGeoCodes(await _geoCodingApiClient.GeoCodeByLocationName(_apiKey, query, limit))
+    return await _geoApi.Call(async api =>
+      {
+        var geoCodes = await api.GeoCodeByLocationName(_apiKey, query, limit);
+        return geoCodes.Select(gc => gc.ToGeoCode());
+      },
+      () => $"GeoCode_{query}_{limit}"
     );
   }
 
@@ -36,20 +39,12 @@ public sealed class GeocodingApi : AbstractApiImplBase, IGeocodingApi
   public async Task<IEnumerable<GeoCode>> QueryReverseAsync(double lat, double lon,
     int limit = int.MaxValue)
   {
-    return await Cached(() => $"GeoCodeReverse_{lat}_{lon}_{limit}",
-      async () => MapGeoCodes(await _geoCodingApiClient.GeoCodeReverse(_apiKey, lat, lon, limit))
+    return await _geoApi.Call(async api =>
+      {
+        var geoCodes = await api.GeoCodeReverse(_apiKey, lat, lon, limit);
+        return geoCodes.Select(gc => gc.ToGeoCode());
+      },
+      () => $"GeoCodeReverse_{lat}_{lon}_{limit}"
     );
-  }
-
-  private static IEnumerable<GeoCode> MapGeoCodes(IApiResponse<ApiGeoCodeResponse[]> response)
-  {
-    if (!response.IsSuccessStatusCode)
-    {
-      throw new ApiException(response.StatusCode, response.ReasonPhrase, response.Error);
-    }
-
-    return response.Content == null
-      ? Enumerable.Empty<GeoCode>()
-      : response.Content.Select(gc => gc.ToGeoCode());
   }
 }