+++ OpenWeatherMap.NetClient/RestApis/Clients/IForecast5DaysApiClient.cs
@@ -15,7 +15,7 @@ internal interface IForecast5DaysApiClient
   /// <param name="lon">Geographical coordinates (longitude)</param>
   /// <param name="cnt">Number of timestamps which will be returned</param>
   [Get("/data/2.5/forecast")]
-  Task<ApiResponse<ApiForecast5DaysResponse>> Forecast(string appid, string lang, double lat, double lon, int cnt);
+  Task<ApiForecast5DaysResponse> Forecast(string appid, string lang, double lat, double lon, int cnt);
   
   /// <summary>
   /// Get the forecast by city ID
@@ -25,5 +25,5 @@ internal interface IForecast5DaysApiClient
   /// <param name="id">City ID</param>
   /// <param name="cnt">Number of timestamps which will be returned</param>
   [Get("/data/2.5/forecast")]
-  Task<ApiResponse<ApiForecast5DaysResponse>> Forecast(string appid, string lang, int id, int cnt);
+  Task<ApiForecast5DaysResponse> Forecast(string appid, string lang, int id, int cnt);
 }