+++ OpenWeatherMap.NetClient/RestApis/Clients/IGeocodingApiClient.cs
@@ -1,5 +1,5 @@
+using OpenWeatherMap.NetClient.RestApis.Responses;
 using Refit;
-using ApiGeoCodeResponse = OpenWeatherMap.NetClient.RestApis.Responses.ApiGeoCodeResponse;
 
 namespace OpenWeatherMap.NetClient.RestApis.Clients;
 
@@ -13,7 +13,7 @@ internal interface IGeocodingApiClient
   /// <param name="q">City name, state code (only for the US) and country code divided by comma (ISO 3166 country codes)</param>
   /// <param name="limit">Number of the locations in the API response (up to 5)</param>
   [Get("/geo/1.0/direct")]
-  Task<ApiResponse<ApiGeoCodeResponse[]>> GeoCodeByLocationName(string appid, string q, int limit);
+  Task<ApiGeoCodeResponse[]> GeoCodeByLocationName(string appid, string q, int limit);
 
   /// <summary>
   /// Get geographical coordinates for a zip/post code
@@ -21,7 +21,7 @@ internal interface IGeocodingApiClient
   /// <param name="appid">API key</param>
   /// <param name="zip">Zip/post code and country code divided by comma (ISO 3166 country codes)</param>
   [Get("/geo/1.0/zip")]
-  Task<ApiResponse<ApiGeoCodeResponse?>> GeoCodeByZipCode(string appid, string zip);
+  Task<ApiGeoCodeResponse?> GeoCodeByZipCode(string appid, string zip);
 
   /// <summary>
   /// Get the name of a location by using geographical coordinates
@@ -31,5 +31,5 @@ internal interface IGeocodingApiClient
   /// <param name="lon">Geographical coordinates (longitude)</param>
   /// <param name="limit">Number of the location names in the API response (several results can be returned in the API response)</param>
   [Get("/geo/1.0/reverse")]
-  Task<ApiResponse<ApiGeoCodeResponse[]>> GeoCodeReverse(string appid, double lat, double lon, int limit);
+  Task<ApiGeoCodeResponse[]> GeoCodeReverse(string appid, double lat, double lon, int limit);
 }